+++ trunk/src/ta3d/src/console/console.api.cpp
@@ -405,8 +405,8 @@ namespace TA3D
 				units.unit[id].Pos.y  = target_pos.y;
 				units.unit[id].cur_px = (sint16)target_pos.x;
 				units.unit[id].cur_py = (sint16)target_pos.z;
-				units.unit[id].Pos.x  = target_pos.x * 8.0f - the_map->map_w_d;
-				units.unit[id].Pos.z  = target_pos.z * 8.0f - the_map->map_h_d;
+				units.unit[id].Pos.x  = target_pos.x * 8.0f - static_cast<float>(the_map->map_w_d);
+				units.unit[id].Pos.z  = target_pos.z * 8.0f - static_cast<float>(the_map->map_h_d);
 				units.unit[id].draw_on_map();
 
 				if (unit_manager.unit_type[units.unit[id].type_id]->ActivateWhenBuilt)// Start activated
@@ -579,7 +579,7 @@ namespace TA3D
 		if (Battle::Instance()->selected) // Sur les unités sélectionnées
 		{
 			units.lock();
-			for (int e = 0; e < units.index_list_size; ++e)
+			for (uint32 e = 0 ; e < units.index_list_size ; ++e)
 			{
 				const int i = units.idx_list[e];
 				if ((units.unit[i].flags & 1) && units.unit[i].owner_id == players.local_human_id && units.unit[i].sel)
