+++ trunk/src/ta3d/src/misc/grid.cpp
@@ -39,7 +39,7 @@ namespace TA3D
 				backup[x] = grid(x, y);
 			for(int x = 0 ; x < grid.getWidth() ; ++x)
 			{
-				float acc((2 * s - (Math::Min(s, grid.getWidth() - 1 - x) - Math::Max(-s, -x))) * 255.0f);
+				float acc(static_cast<float>((2 * s - (Math::Min(s, grid.getWidth() - 1 - x) - Math::Max(-s, -x))) * 255));
 				for(int i = Math::Max(-s, -x) ; i <= Math::Min(s, grid.getWidth() - 1 - x) ; ++i)
 					acc += kernel[i + s] * backup[x + i];
 				grid(x,y) = acc;
@@ -60,7 +60,7 @@ namespace TA3D
 				backup[y] = grid(x, y);
 			for(int y = 0 ; y < grid.getHeight() ; ++y)
 			{
-				float acc((2 * s - (Math::Min(s, grid.getHeight() - 1 - x) - Math::Max(-s, -y))) * 255.0f);
+				float acc(static_cast<float>((2 * s - (Math::Min(s, grid.getHeight() - 1 - x) - Math::Max(-s, -y))) * 255));
 				for(int i = Math::Max(-s, -y) ; i <= Math::Min(s, grid.getHeight() - 1 - y) ; ++i)
 					acc += kernel[i + s] * backup[y + i];
 				grid(x,y) = acc;
